This function returns a data frame shows player information about a specifical match when passed match id

This function returns a data frame shows player information about a specifical match when passed match id

match_details(matchID)

Arguments

matchID
a dota2 match ID

Examples

match_details(4172245371)

#>          hero name   player_name account_id    team levels kills deaths assists
#> 83       Ogre Magi          MaTa  298726580 Radiant     15     2      9       7
#> 70  Spirit Breaker       Slander  129340473 Radiant     14     1     12      13
#> 10       Morphling      BarabaNN   91380904 Radiant     22     7      3       5
#> 17    Storm Spirit          june  196012404 Radiant     22     5      7       3
#> 44           Pugna            .D   88655279 Radiant     18     3      8       8
#> 31            Riki BabySchmidty-  105248644    Dire     24    17      2      13
#> 77      Brewmaster        AstAzE  395558912    Dire     19     5      5      20
#> 111  Winter Wyvern          Eixn  116876282    Dire     16     2      4      17
#> 99            Tusk            gh  101356886    Dire     18     4      5      27
#> 108    Terrorblade        Rajjix   88826062    Dire     22     9      2      15
#>     kda last_hits denies gold_per_min xp_per_min hero_damage tower_damage
#> 83    0        10      2          236        308        7914            0
#> 70    1        22      1          198        250        9383            0
#> 10    3       293     25          536        564       16783         1775
#> 17    1       288      8          519        565       15758           94
#> 44    1       152      6          371        395       14717         2506
#> 31   10       178     32          551        688       28752         1594
#> 77    4       116      8          377        440       10057          723
#> 111   3        53      0          343        348       11371          239
#> 99    5        36      2          345        418       13198          182
#> 108   8       331     10          689        573       19180        11701