This function returns a data frame shows player information about a specifical match when passed a match id

This function returns a data frame shows player information about a specifical match when passed a match id

match_details(matchID)

Arguments

matchID
a dota2 match ID

Examples

match_details(4172245371)

#>          hero name   player_name account_id    team status levels kills deaths
#> 83       Ogre Magi          MaTa  298726580 Radiant   lose     15     2      9
#> 70  Spirit Breaker       Slander  129340473 Radiant   lose     14     1     12
#> 10       Morphling      BarabaNN   91380904 Radiant   lose     22     7      3
#> 17    Storm Spirit          june  196012404 Radiant   lose     22     5      7
#> 44           Pugna            .D   88655279 Radiant   lose     18     3      8
#> 31            Riki BabySchmidty-  105248644    Dire    win     24    17      2
#> 77      Brewmaster        AstAzE  395558912    Dire    win     19     5      5
#> 111  Winter Wyvern          Eixn  116876282    Dire    win     16     2      4
#> 99            Tusk            gh  101356886    Dire    win     18     4      5
#> 108    Terrorblade        Rajjix   88826062    Dire    win     22     9      2
#>     assists kda last_hits denies gold_per_min total_gold xp_per_min total_xp
#> 83        7   0        10      2          236       8527        308    11129
#> 70       13   1        22      1          198       7154        250     9033
#> 10        5   3       293     25          536      19367        564    20379
#> 17        3   1       288      8          519      18753        565    20415
#> 44        8   1       152      6          371      13405        395    14272
#> 31       13  10       178     32          551      19909        688    24859
#> 77       20   4       116      8          377      13622        440    15898
#> 111      17   3        53      0          343      12393        348    12574
#> 99       27   5        36      2          345      12466        418    15103
#> 108      15   8       331     10          689      24895        573    20704
#>     hero_damage tower_damage hero_heal
#> 83         7914            0         0
#> 70         9383            0         0
#> 10        16783         1775         0
#> 17        15758           94         0
#> 44        14717         2506       444
#> 31        28752         1594         0
#> 77        10057          723      1066
#> 111       11371          239      1076
#> 99        13198          182      1683
#> 108       19180        11701         0